spidev_device now is unit16_t

+++ lib/include/libsoc_spi.h
@@ -53,7 +53,7 @@ typedef enum {
  * \param uint8_t spidev_device - the minor spidev number
  * \return spi* struct pointer or NULL on failure
  */
-spi* libsoc_spi_init (uint8_t spidev_device, uint8_t chip_select);
+spi* libsoc_spi_init (uint16_t spidev_device, uint8_t chip_select);
 
 /**
  * \fn int libsoc_spi_free(spi* spi)
